pyproject.toml
[project]
name = "strideai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.6",
    "langchain-openai>=1.1.7",
    "opik>=1.9.89",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
]
